feat: remove exchange modal from home element
