+++ src/components/Home.vue
@@ -4,260 +4,33 @@
     <el-main style="width: 100%; height: 100vh; padding: 0; padding-left: 62px;">
       <router-view />
     </el-main>
-    <el-dialog
-      title="Exchange"
-      width="450px"
-      top="2vh"
-      :visible="exchangeDialogVisible"
-      @close="closeExchangeDialogWith()"
-      center
-    >
-      <el-form ref="exchangeForm" :model="exchangeForm" class="exchange_form" :rules="rules">
-        <el-form-item label="I send" prop="offer_amount" >
-          <el-input name="amount" v-model="exchangeForm.offer_amount" placeholder="0">
-            <el-select
-              v-model="exchangeDialogOfferAsset"
-              @change="getOfferToRequestPrice()"
-              slot="append"
-              placeholder="asset"
-              style="width: 100px"
-            >
-              <el-option
-                v-for="wallet in assetsWithoutRequest"
-                :key="wallet.id"
-                :label="wallet.asset"
-                :value="wallet.asset">
-                  <span style="float: left">{{ `${wallet.name} (${wallet.asset})` }}</span>
-              </el-option>
-            </el-select>
-          </el-input>
-        </el-form-item>
-        <span class="form-item-text">
-          Available balance:
-          <span v-if="exchangeDialogOfferAsset" class="form-item-text-amount">
-            {{ wallets.find(x => x.asset === exchangeDialogOfferAsset).amount | formatPrecision }} {{ exchangeDialogOfferAsset }}
-          </span>
-          <span v-else>...</span>
-        </span>
-        <el-form-item label="I receive" prop="request_amount">
-          <el-input name="amount" v-model="exchangeForm.request_amount" placeholder="0">
-            <el-select
-              v-model="exchangeDialogRequestAsset"
-              @change="getOfferToRequestPrice()"
-              slot="append"
-              placeholder="asset"
-              style="width: 100px"
-            >
-              <el-option
-                v-for="wallet in assetsWithoutOffer"
-                :key="wallet.id"
-                :label="wallet.asset"
-                :value="wallet.asset">
-                  <span style="float: left">{{ `${wallet.name} (${wallet.asset})` }}</span>
-              </el-option>
-            </el-select>
-          </el-input>
-        </el-form-item>
-        <span class="form-item-text">
-          Market price:
-          <span v-if="exchangeDialogRequestAsset && exchangeDialogOfferAsset && exchangeDialogPrice" class="form-item-text-amount">
-            1 {{ exchangeDialogOfferAsset }} ≈ {{ exchangeDialogPrice }} {{ exchangeDialogRequestAsset }}
-          </span>
-          <span v-else>...</span>
-        </span>
-        <el-form-item label="Counterparty" prop="to">
-          <el-input v-model="exchangeForm.to" placeholder="Account id"/>
-        </el-form-item>
-        <el-form-item label="Additional information" prop="description">
-          <el-input
-            v-model="exchangeForm.description"
-            placeholder="Description"
-          />
-        </el-form-item>
-      </el-form>
-      <el-button
-        class="dialog-form_buttons action fullwidth"
-        @click="onSubmitExchangeDialog()"
-        style="margin-top: 40px"
-        :loading="isExchangeSending"
-      >
-        EXCHANGE
-      </el-button>
-    </el-dialog>
-    <confirm-modal :isExchangeDialogVisible="this.exchangeDialogVisible"/>
+    <exchange-modal />
+    <confirm-modal />
   </el-container>
 </template>
 
 <script>
-import { mapState, mapGetters, mapActions } from 'vuex'
+import { mapGetters } from 'vuex'
 import { lazyComponent } from '@router'
-import inputValidation from '@/components/mixins/inputValidation'
-import numberFormat from '@/components/mixins/numberFormat'
-import messageMixin from '@/components/mixins/message'
-import NOTIFICATIONS from '@/data/notifications'
 
-// TODO: Validate lack of selected asset
 export default {
   name: 'Home',
-  mixins: [
-    messageMixin,
-    numberFormat,
-    inputValidation({
-      to: 'checkAccountId',
-      request_amount: 'tokensAmount',
-      offer_amount: 'tokensAmount',
-      description: 'additionalInformation'
-    })
-  ],
   components: {
     Menu: lazyComponent('Home/Menu'),
-    ConfirmModal: lazyComponent('common/modals/ConfirmModal')
+    ConfirmModal: lazyComponent('common/modals/ConfirmModal'),
+    ExchangeModal: lazyComponent('common/modals/ExchangeModal')
   },
   data () {
-    return {
-      exchangeForm: {
-        to: null,
-        request_amount: '',
-        offer_amount: '',
-        description: null
-      },
-      isExchangeSending: false
-    }
+    return {}
   },
-
-  computed: {
-    ...mapGetters([
-      'wallets',
-      'exchangeDialogVisible',
-      'exchangeDialogPrice',
-      'accountQuorum'
-    ]),
-
-    ...mapState({
-      accountId: (state) => state.Account.accountId
-    }),
-
-    assetsWithoutOffer () {
-      return this.wallets.filter(a => a.asset !== this.exchangeDialogOfferAsset)
-    },
-
-    assetsWithoutRequest () {
-      return this.wallets.filter(a => a.asset !== this.exchangeDialogRequestAsset)
-    },
-
-    exchangeDialogOfferAsset: {
-      get () {
-        return this.$store.getters.exchangeDialogOfferAsset
-      },
-      set (asset) {
-        this.$store.commit('SET_EXCHANGE_DIALOG_OFFER_ASSET', asset)
-      }
-    },
-
-    exchangeDialogRequestAsset: {
-      get () {
-        return this.$store.getters.exchangeDialogRequestAsset
-      },
-      set (asset) {
-        this.$store.commit('SET_EXCHANGE_DIALOG_REQUEST_ASSET', asset)
-      }
-    },
-
-    numberOfSettlements () {
-      return this.$store.getters.incomingSettlements.length
-    }
-  },
-
   created () {
     this.$store.dispatch('getAllUnsignedTransactions')
     this.$store.dispatch('loadSettings')
   },
-
-  beforeUpdate () {
-    const wallet = this.wallets.find(x => x.asset === this.exchangeDialogOfferAsset)
-    const { amount, precision } = wallet || { amount: 0, precision: 0 }
-    this._refreshRules({
-      offer_amount: {
-        pattern: 'tokensAmount',
-        amount: amount,
-        precision: precision,
-        asset: this.exchangeDialogOfferAsset
-      },
-      request_amount: {
-        pattern: 'tokensAmount',
-        amount: Number.MAX_SAFE_INTEGER,
-        precision: precision,
-        asset: this.exchangeDialogRequestAsset
-      },
-      to: { pattern: 'checkAccountId', accountId: this.exchangeForm.to }
-    })
-  },
-
-  updated () {
-    if (this.$refs.exchangeForm && !this.exchangeDialogVisible) {
-      this.$refs.exchangeForm.resetFields()
-    }
-  },
-
-  methods: {
-    ...mapActions([
-      'openApprovalDialog',
-      'closeExchangeDialog',
-      'getOfferToRequestPrice'
-    ]),
-
-    closeExchangeDialogWith () {
-      this.closeExchangeDialog()
-      this.exchangeForm.description = ''
-      this.exchangeDialogOfferAsset = ''
-      this.exchangeDialogRequestAsset = ''
-    },
-
-    onSubmitExchangeDialog () {
-      const s = this.exchangeForm
-      this.$refs.exchangeForm.validate(valid => {
-        if (!valid) return
-
-        this.openApprovalDialog()
-          .then(privateKeys => {
-            if (!privateKeys) return
-            this.isExchangeSending = true
-            const offerAsset = this.wallets.find(x => x.asset === this.exchangeDialogOfferAsset).assetId
-            const requestAsset = this.wallets.find(x => x.asset === this.exchangeDialogRequestAsset).assetId
-            return this.$store.dispatch('createSettlement', {
-              privateKeys,
-              to: s.to,
-              offerAssetId: offerAsset,
-              offerAmount: s.offer_amount,
-              requestAssetId: requestAsset,
-              requestAmount: s.request_amount,
-              description: s.description
-            })
-              .then(() => {
-                let completed = privateKeys.length === this.accountQuorum
-                this.$_showMessageFromStatus(
-                  completed,
-                  NOTIFICATIONS.SETTLEMENT_SUCCESS,
-                  NOTIFICATIONS.NOT_COMPLETED
-                )
-
-                this.closeExchangeDialogWith()
-                // TODO: think, maybe it is a bad idea to close form after success.
-                Object.assign(
-                  this.$data.exchangeForm,
-                  this.$options.data().exchangeForm
-                )
-              })
-              .catch(err => {
-                console.error(err)
-                this.$_showErrorAlertMessage(err.message, 'Withdrawal error')
-              })
-              .finally(() => {
-                this.isExchangeSending = false
-              })
-          })
-      })
-    }
+  computed: {
+    ...mapGetters([
+      'accountQuorum'
+    ])
   }
 }
 </script>
