+++ src/components/common/modals/ConfirmModal.vue
@@ -18,7 +18,7 @@
         <el-row class="approval_form-desc">
           <p>
             Please enter your private key<span v-if="accountQuorum > 1">s</span>.
-            <span v-if="accountQuorum > 1 && !isExchangeDialogVisible">
+            <span v-if="accountQuorum > 1 && !exchangeDialogVisible">
               You need to enter at least {{ approvalDialogMinAmountKeys }} key.
             </span>
           </p>
@@ -139,12 +139,6 @@ export default {
   mixins: [
     errorHandler
   ],
-  props: {
-    isExchangeDialogVisible: {
-      type: Boolean,
-      required: true
-    }
-  },
   validations () {
     return {
       approvalForm: {
@@ -173,6 +167,7 @@ export default {
   },
   computed: {
     ...mapGetters([
+      'exchangeDialogVisible',
       'approvalDialogVisible',
       'approvalDialogSignatures',
       'approvalDialogMinAmountKeys',
@@ -250,7 +245,7 @@ export default {
       this.$set(this.approvalForm, 'numberOfValidKeys', tempNumberOfValidKeys)
     },
     disableConfig () {
-      if (this.isExchangeDialogVisible) {
+      if (this.exchangeDialogVisible) {
         return !(this.approvalForm.numberOfValidKeys + this.approvalDialogSignatures.length === this.accountQuorum)
       } else {
         if (this.approvalDialogMinAmountKeys === 1) {
